+++ sdk/typespec/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.9.0 (Unreleased)
+## 0.9.0 (2025-10-03)
 
 ### Features Added
 
@@ -26,8 +26,6 @@
   - Renamed `ResultExt::map_kind()` to `ResultExt::with_kind()`.
   - Renamed `ResultExt::with_context()` to `ResultExt::with_context_fn()`.
   - Renamed `ResultExt::context()` to `ResultExt::with_context()`.
-
-### Bugs Fixed
 
 ### Other Changes